feat: 최신 일기 데이터 유지
- 최신 일기 데이터을 바로 반영하기 위해 상위 컴포넌트 MainPage에서 데이터 관리
- 모달 상태 변경 때마다 일기 데이터 최신화
